requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
fastapi>=0.115.0
uvicorn>=0.31.0
pydantic>=2.9.0
python-dotenv>=1.0.0
openai>=1.51.0
tiktoken>=0.8.0
pypdf>=5.0.0
